pyproject.toml
[tool.poetry]
name = "textbase-client"
version = "0.1.2"
description = "A python library to help you create chatbots easily with some more configurations"
authors = ["kaus-cofactory <kaustubhshukla@cofactory.ai>", "sammyCofactory <samarthar@cofactory.ai>"]
readme = "README.md"
packages = [{include = "textbase"}]

[tool.poetry.dependencies]
python = "^3.9.0"
requests = "^2.31.0"
openai = "^0.27.8"
python-dotenv = "^1.0.0"
tabulate = "^0.9.0"
functions-framework = "^3.4.0"
yaspin = "^3.0.0"
pydantic = "^1.9"
langchain = "^0.0.279"
chromadb = "^0.4.8"
tiktoken = "^0.4.0"
Cython = "0.29.36"
pdf2image = "^1.16.3"
unstructured = "^0.10.11"
twilio = "^8.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
textbase-client = "textbase.textbase_cli:cli"

